+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add facility to merge `Serializers` instances via `Serializers.merge`,
   `SerializersBuilder.merge` and `SerializersBuilder.mergeAll`.
+- Bug fix: fix generated code with polymorphism and more than one level of
+  non-instantiable classes.
 - Bug fix: fix generated `operator==` when a field is called `other`.
 - Fix `num` deserialization so it does not always convert to `double`.
 - Bump version of `analyzer`.